- 자동로그인 판별을 Splash Activity에서 하도록 하였습니다. - 릴리즈 버전에서 구글맵 API키를 불러오지 못해 API키 저장하는 방식을 변경하였습니다.
