+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
 
         <meta-data
             android:name="com.google.android.geo.API_KEY"
-            android:value="@string/google_maps_key"/>
+            android:value="${googleMapApiKey}"/>
 
         <activity
             android:name=".activity.SplashActivity"
