+++ gradle.properties
@@ -22,4 +22,3 @@ kotlin.code.style=official
 # thereby reducing the size of the R class for that library
 android.nonTransitiveRClass=true
 android.enableJetifier=true
-GOOGLE_MAPS_API_KEY=AIzaSyA_ppNk6FfxhQVmU59Ic5zmaS_9KEbeipk
